% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST33.R
\docType{data}
\name{ST33}
\alias{ST33}
\title{MTBLS20-ST439033 dataset}
\format{
An object of class list of length 2.
}
\source{
Extracted from MetaboLights (accession number: MTBLS20,
sub-study: 439033)
}
\usage{
data(ST33)
}
\description{
This study was aimed at evaluating the impact of collection and
short-term storage conditions on the metabolome of human urine samples.
The identifications reported by this study were achieved by at
least comparing two physicochemical parameters to those of a
reference compound (level 2 identification).
It consists of a list containing
the positive (\code{Positive}) acquisition mode data and the
negative (\code{Negative}) acquisition mode data. Within each
element, a list of length 2 is available, which contains the
input peak list (\code{Input}) and the reference peak to compound
assignations (\code{Output}).
}
\references{
Roux,  A. et al.
\emph{Annotation of the Human Adult Urinary Metabolome and
Metabolite Identification Using Ultra High Performance Liquid
Chromatography Coupled to a Linear Quadrupole Ion Trap-Orbitrap
Mass Spectrometer}, Analytical chemistry 2012
}
\keyword{datasets}
