% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trypanosoma.R
\docType{data}
\name{Trypanosoma}
\alias{Trypanosoma}
\title{Trypanosoma brucei dataset}
\format{
An object of class list of length 2.
}
\source{
Extracted from the ProbMetab website.
}
\usage{
data(Trypanosoma)
}
\description{
This study contains a set of compounds identified with the aid of standard
references (level 1 identification), which makes it specially
suited for validation purposes. It consists of a list containing
the positive (\code{Positive}) acquisition mode data and the
negative (\code{Negative}) acquisition mode data. Within each
element, a list of length 2 is available, which contains the
input peak list (\code{Input}) and the reference peak to compound
assignations (\code{Output}).
}
\references{
Silva,  R. et al.
\emph{ProbMetab: An R package for Bayesian prob-abilistic
annotation  of  LC-MS-based  metabolomics}, Bioinformatics 2014
}
\keyword{datasets}
